# [TechDebt: Issue #260] Plugin execution improvements