@dikhan 2020-07-20
Using integration tests to benchmark the improvements (even though this may not necessarily represent the actual improvement due to the simplicity of the integration tests simple OpenAPI document):

- Before

````
$ make integration-test  10.19s user 4.23s system 41% cpu 35.041 total
````

- After

````
$ make integration-test  10.01s user 4.25s system 41% cpu 34.506 total
````